refactor: CollectionFilmUiModel 파일 분리 및 상태 관리 개선
- `CollectionCreateScreen` 내에 있던 `CollectionFilmUiModel`을 별도의 파일로 분리했습니다.
- filmList의 상태를 `ImmutableList`로 관리하고, 삭제 로직을 상위 컴포저블로 이동하여 상태 관리 방식을 개선했습니다.
- `CollectionCreateFilmItemList`의 취소 아이콘 색상을 `Color.White`로 수정하고, 플레이스홀더 텍스트를 "이 작품의 매력 포인트를 적어주세요."로 변경했습니다.
